requirements.txt
python-dotenv==0.14.0
discord.py==1.5.0
